Automatically buy augments and then reset

+++ src/destroy.ts
@@ -1,6 +1,14 @@
 import { NS } from '@ns';
 
+const WORLD_DAEMON = 'w0rld_d43m0n';
 
 export async function main(ns: NS): Promise<void> {
-    ns.singularity.destroyW0r1dD43m0n(4, 'init.ts');
+    if (ns.serverExists(WORLD_DAEMON) && ns.hasRootAccess(WORLD_DAEMON)) {
+
+
+
+        ns.getBitNodeMultipliers()
+        ns.singularity.destroyW0r1dD43m0n(4, 'init.ts');
+
+    }
 }
